Crate atspi

atspi root crate, which may bring in various subcrates:

atspi_common (required)

atspi_proxies (with use of proxies feature flag)

atspi_connection (with use of connection feature flag)

zbus re-export (with use of zbus feature flag)

Re-exports

pub use atspi_proxies as proxy;

Modules

action

cache

Common types for org.a11y.atspi.Cache events.

error

events

interface

Conversion functions and types representing a set of Interfaces.

object_match

object_ref

operation

state

Structs

Action

An action which may be triggered through the accessibility API.

CacheItem

The item type provided by Cache:Add signals

InterfaceSet

A collection type which encodes the AT-SPI interfaces an accessible object has implemented.

LegacyCacheItem

The item type provided by Cache:Add signals

ObjectMatchRule

Definition of match rules for accessible objects. Rule(s) against which we can match an object or a collection thereof.

ObjectRefOwned

A wrapper around the static variant of ObjectRef. This is guaranteed to have a 'static lifetime.

StateSet

The bitflag representation of all states an object may have.

TextSelection

Describes a selection of text, including selections across object boundaries.

Enums

AtspiError

An error type that can describe atspi and std and different zbus errors.

CacheEvents

All events related to the org.a11y.atspi.Cache interface. Note that these are not telling the client that an item has been added to a cache. It is telling the client “here is a bunch of information to store it in your cache”.

ClipType

Enumeration used by TextProxy to indicate how to treat characters intersecting bounding boxes.

CoordType

The coordinate type encodes the frame of reference.

DocumentEvents

Event

Encapsulates the various different accessibility bus signal types.

EventListenerEvents

The events that can be emitted by the registry daemon. This enum is used to wrap the events that are emitted by the registry daemon. The events are EventListenerRegisteredEvent and EventListenerDeregisteredEvent.

FocusEvents

Granularity

Level of granularity to get text of, in relation to a cursor position.

Interface

AT-SPI interfaces an accessible object can implement.

KeyboardEvents

Layer

Indicates relative stacking order of a atspi_proxies::component::ComponentProxy with respect to the onscreen visual representation of the UI.

MatchType

Enumeration used by ObjectMatchRule to specify how to interpret ObjectRef objects.

MouseEvents

ObjectEvents

ObjectRef

A unique identifier for an object in the accessibility tree.

Operation

An operation can either be Self::Insert or Self::Delete. These correspond to methods available on Vec.

Politeness

Enumeration used to indicate a type of live region and how assertive it should be in terms of speaking notifications.

RelationType

Describes a relationship between one object and another.

Role

An accessible object role.

ScrollType

Enumeration used by interface the crate::interface::Interface::Accessible to specify where an object should be placed on the screen when using scroll_to.

SortOrder

Enumeration used by interface CollectionProxy to specify the way ObjectRef objects should be sorted.

State

Used by various interfaces indicating every possible state of an accessibility object.

TerminalEvents

All events related to the org.a11y.atspi.Event.Terminal interface.

TreeTraversalType

Defines how an object-tree is to be traversed. Used in CollectionProxy.

WindowEvents

All events on the org.a11y.atspi.Event.Window interface.

Traits

EventProperties

EventProperties allows access to the internals of an event, specifically:

EventTypeProperties

Describes properties of a specific event type.

Type Aliases

Result