Module atspi::component

DBus interface proxy for: org.a11y.atspi.Component

This code was generated by zbus-xmlgen 2.0.1 from DBus introspection data. Source: Component.xml.

You may prefer to adapt it, instead of using it verbatim.

More information can be found in the Writing a client proxy section of the zbus documentation.

Structs

ComponentProxy

ComponentProxyBlocking

Enums

Layer

Indicates relative stacking order of a ComponentProxy with respect to the onscreen visual representation of the UI.

ScrollType

Traits

Component

ComponentBlocking