atspi::events

Module document

Structs

AttributesChangedEvent

ContentChangedEvent

LoadCompleteEvent

An event triggered by the completion of a document load action. For example: a web page has finished loading its initial payload, or LibreOffice has loaded a document from disk.

LoadStoppedEvent

An event triggered by the cancelling of a document load. For example: during the loading of a large web page, a user may press Escape to stop loading the page.

PageChangedEvent

ReloadEvent

An event triggered by a reloading of a document. For example: pressing F5, or Control + r will reload a page in a web browser.

Enums

DocumentEvents