atspi::events::event_body

Struct EventBodyBorrowed

pub struct EventBodyBorrowed<'a> {
    pub kind: &'a str,
    pub detail1: i32,
    pub detail2: i32,
    pub any_data: Value<'a>,
    /* private fields */
}

Fields
kind: &'a str

kind variant, used for specifying an event triple โobject:state-changed:focusedโ, the โfocusโ part of this event is what is contained within the kind.

detail1: i32

Generic detail1 value described by AT-SPI.

detail2: i32

Generic detail2 value described by AT-SPI.

any_data: Value<'a>

Generic any_data value described by AT-SPI. This can be any type.

Implementations

impl EventBodyBorrowed<'_>

pub fn to_fully_owned(&self) -> Result<EventBodyOwned, AtspiError>

Convert this borrowed event body to an owned event body.

Errors

This will error if the following conditions are met:

the any_data field contains an std::os::fd::OwnedFd type, and

the maximum number of open files for the process is exceeded.

Chances are slim because none of the types in crate::events use std::os::fd::OwnedFd. See zvariant::Value::try_clone for more information.