Struct atspi::events::AtspiEvent

pub struct AtspiEvent { /* private fields */ }

Implementations

impl AtspiEvent

pub fn body(&self) -> &EventBodyOwned

Deserialized signal body type.

pub fn event_string(&self) -> String

Returns the atspi event string for this event type (E.G. “Object:StateChanged:Focused”).

This should not be used for matching on events as it needlessly allocates and copies the 3 components of the event type. It is meant for logging, etc.

pub fn kind(&self) -> &str

pub fn detail1(&self) -> i32

Event dependant detail.

pub fn detail2(&self) -> i32

Event dependant detail.

pub fn any_data(&self) -> &OwnedValue

Event dependant generic Value.

pub fn properties(&self) -> &HashMap<String, OwnedValue>