atspi::events

Trait MessageConversion

pub trait MessageConversion<'a>: DBusProperties {
    type Body<'msg>: Type + Deserialize<'msg> + Serialize
       where Self: 'msg;

    // Required methods
    fn from_message_unchecked(
        msg: &Message,
        header: &Header<'_>,
    ) -> Result<Self, AtspiError>
       where Self: Sized + 'a;
    fn from_message_unchecked_parts(
        obj_ref: ObjectRef<'_>,
        body: Body,
    ) -> Result<Self, AtspiError>
       where Self: Sized;
    fn body(&self) -> Self::Body<'_>;
}

Required Associated Types

type Body<'msg>: Type + Deserialize<'msg> + Serialize where Self: 'msg

What is the body type of this event.

Required Methods

fn from_message_unchecked( msg: &Message, header: &Header<'_>, ) -> Result<Self, AtspiError>
where Self: Sized + 'a,

Build an event from a zbus::Message reference. This function will not check for any of the following error conditions:

That the message has an interface: AtspiError::MissingInterface

That the message interface matches the one for the event: AtspiError::InterfaceMatch

That the message has an member: AtspiError::MissingMember

That the message member matches the one for the event: AtspiError::MemberMatch

That the message signature matches the one for the event: AtspiError::SignatureMatch

Therefore, this should only be used when one has checked the above conditions. These must be checked manually. Alternatively, there is the MessageConversionExt::try_from_message that will check these conditions for you.

This type also implements TryFrom<&zbus::Message>; consider using this if you are not an internal developer.

Errors

It is possible to get a AtspiError::Zvariant error if you do not check the proper conditions before calling this.

fn from_message_unchecked_parts( obj_ref: ObjectRef<'_>, body: Body, ) -> Result<Self, AtspiError>
where Self: Sized,

Build an event from an ObjectRef and Self::Body. This function will not check for any of the following error conditions:

That the message has an interface: AtspiError::MissingInterface

That the message interface matches the one for the event: AtspiError::InterfaceMatch

That the message has an member: AtspiError::MissingMember

That the message member matches the one for the event: AtspiError::MemberMatch

Therefore, this should only be used when one has checked the above conditions.

Errors

Some Self::Body types may fallibly convert data fields contained in the body. If this happens, then the function will return an error.

fn body(&self) -> Self::Body<'_>

The body of the object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.