atspi::events

Trait RegistryEventString

pub trait RegistryEventString {
    const REGISTRY_EVENT_STRING: &'static str;
}

A static Registry event string for registering with the RegistryProxy for receiving events.

Required Associated Constants

const REGISTRY_EVENT_STRING: &'static str

A registry event string for registering for event receiving via the RegistryProxy. This should be deprecated in favour of composing the string from DBusMember::DBUS_MEMBER and DBusInterface::DBUS_INTERFACE.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.