Module atspi::identify

Signified signal types

The generic AtspiEvent has a specific meaning depending on its origin. This module offers the signified types and their conversions from a generic AtpiEvent.

The TrySignify macro implements a TryFrom<Event> on a per-name and member basis

Modules

object

Structs

AbsEvent

ActivateEvent

ApplicationChangedEvent

AttributesChangedEvent

ButtonEvent

CharwidthChangedEvent

CloseEvent

ColumncountChangedEvent

ContentChangedEvent

CreateEvent

DeactivateEvent

DesktopCreateEvent

DesktopDestroyEvent

DestroyEvent

FocusEvent

LineChangedEvent

LinecountChangedEvent

LoadCompleteEvent

LoadStoppedEvent

LowerEvent

MaximizeEvent

MinimizeEvent

ModifiersEvent

MoveEvent

PageChangedEvent

PropertyChangeEvent

RaiseEvent

RelEvent

ReloadEvent

ReparentEvent

ResizeEvent

RestoreEvent

RestyleEvent

ShadeEvent

uUshadeEvent

Enums

DocumentEvents

Any of the Document events.

FocusEvents

The Focus event.

KeyboardEvents

The Keyboard events.

MouseEvents

Any of the Mouse events.

ObjectEvents

Any of the Object events.

TerminalEvents

Any of the Terminal events.

WindowEvents

Any of the Window events.

Traits

Signified

All Atspi / Qspi event types encapsulate AtspiEvent. This trait allows access to the underlying item.