atspi::object_ref

Enum ObjectRef

pub enum ObjectRef<'o> {
    Null,
    Owned {
        name: UniqueName<'static>,
        path: ObjectPath<'static>,
    },
    Borrowed {
        name: UniqueName<'o>,
        path: ObjectPath<'o>,
    },
}

A unique identifier for an object in the accessibility tree.

A ubiquitous type used to refer to an object in the accessibility tree.

In AT-SPI2, objects in the applications’ UI object tree are uniquely identified using an application’s bus name and object path. “(so)”

Emitted by RemoveAccessible and Available

Variants

Null

Owned

Fields

name: UniqueName<'static>

path: ObjectPath<'static>

Borrowed

Fields

name: UniqueName<'o>

path: ObjectPath<'o>

Implementations

impl<'o> ObjectRef<'o>

pub fn new(name: UniqueName<'o>, path: ObjectPath<'o>) -> ObjectRef<'o>

Create a new ObjectRef::Borrowed from a UniqueName and ObjectPath.

pub fn new_owned<N, P>(name: N, path: P) -> ObjectRefOwned
where N: Into<UniqueName<'static>>, P: Into<ObjectPath<'static>>,

Create a new, owned ObjectRef.

Example

use zbus::names::UniqueName;
use zbus::zvariant::ObjectPath;
use atspi_common::ObjectRef;

let name = UniqueName::from_static_str_unchecked(":1.23");
let path = ObjectPath::from_static_str_unchecked("/org/a11y/example/path/007");

let object_ref = ObjectRef::new_owned(name, path);

pub fn new_borrowed<N, P>(name: N, path: P) -> ObjectRef<'o>
where N: Into<UniqueName<'o>>, P: Into<ObjectPath<'o>>,

Create a new, borrowed ObjectRef.

Example

use zbus::names::UniqueName;
use zbus::zvariant::ObjectPath;
use atspi_common::ObjectRef;

let name = UniqueName::from_static_str_unchecked(":1.23");
let path = ObjectPath::from_static_str_unchecked("/org/a11y/example/path/007");

let object_ref = ObjectRef::new_borrowed(name, path);

pub fn try_from_bus_name_and_path( sender: BusName<'o>, path: ObjectPath<'o>, ) -> Result<ObjectRef<'o>, AtspiError>

Create a new ObjectRef, from BusName and ObjectPath.

Errors

Will fail if the sender is not a UniqueName.

pub const fn from_static_str_unchecked( name: &'static str, path: &'static str, ) -> ObjectRef<'o>

Create a new ObjectRef, unchecked.

Safety

The caller must ensure that the strings are valid.

pub fn is_null(&self) -> bool

Returns true if the object reference is Null, otherwise returns false.

Toolkits may use the Null object reference to indicate that an object is not available or does not exist. For example, when calling Accessible::get_parent on an object that has no parent, it may return a Null object reference.

pub fn name(&self) -> Option<&UniqueName<'o>>

Returns the name of the object reference. If the object reference is Null, it returns None. If the object reference is Owned or Borrowed, it returns the name.

Example

use zbus::names::UniqueName;
use zbus::zvariant::ObjectPath;
use atspi_common::ObjectRef;

let name = UniqueName::from_static_str_unchecked(":1.23");
let path = ObjectPath::from_static_str_unchecked("/org/a11y/example/path/007");
let object_ref = ObjectRef::new_borrowed(name, path);

// Check the name of the object reference
assert!(object_ref.name().is_some());
assert_eq!(object_ref.name().unwrap().as_str(), ":1.23");

pub fn path(&self) -> &ObjectPath<'o>

Returns the path of the object reference.\

Example

use zbus::names::UniqueName;
use zbus::zvariant::ObjectPath;
use atspi_common::ObjectRef;

let name = UniqueName::from_static_str_unchecked(":1.23");
let path = ObjectPath::from_static_str_unchecked("/org/a11y/example/path/007");
let object_ref = ObjectRef::new_borrowed(name, path);

// Check the path of the object reference
assert_eq!(object_ref.path().as_str(), "/org/a11y/example/path/007");

pub fn into_owned(self) -> ObjectRef<'static>

Converts the ObjectRef into an owned instance, consuming self.
If the object reference is Null, it returns ObjectRef::Null.
If the object reference is Owned, it returns the same ObjectRef::Owned.
If the object reference is Borrowed, it converts the name and path to owned versions and returns ObjectRef::Owned.

Extending lifetime ‘magic’ (from ’o -> ‘static’)

ObjectRef<'_> leans on the implementation of UniqueName and ObjectPath to convert the inner types to 'static. These types have an Inner enum that can contain an Owned, Borrowed, or Static Str type. The Strtype is either a &'static str (static), &str (borrowed), or an Arc<str> (owned).

Example

use zbus::names::UniqueName;
use zbus::zvariant::ObjectPath;
use atspi_common::ObjectRef;

let name = UniqueName::from_static_str_unchecked(":1.23");
let path = ObjectPath::from_static_str_unchecked("/org/a11y/example/path/007");
let object_ref = ObjectRef::new_borrowed(name, path);

// Check whether the object reference can be converted to an owned version
assert!(!object_ref.is_null());
let object_ref = object_ref.into_owned();
assert!(matches!(object_ref, ObjectRef::Owned { .. }));

pub fn name_as_str(&self) -> Option<&str>

Returns the name of the object reference as a string slice.

pub fn path_as_str(&self) -> &str

Returns the path of the object reference as a string slice.