Module atspi::registry

DBus interface proxy for: org.a11y.atspi.Registry

This code was generated by zbus-xmlgen 2.0.1 from DBus introspection data. Source: Registry.xml.

You may prefer to adapt it, instead of using it verbatim.

More information can be found in the Writing a client proxy section of the zbus documentation.

Structs

EventListenerDeregistered

A EventListenerDeregistered signal.

EventListenerDeregisteredArgs

EventListenerDeregistered signal arguments.

EventListenerDeregisteredIterator

A Iterator implementation that yields EventListenerDeregistered signals.

EventListenerDeregisteredStream

A Stream implementation that yields EventListenerDeregistered signals.

EventListenerRegistered

A EventListenerRegistered signal.

EventListenerRegisteredArgs

EventListenerRegistered signal arguments.

EventListenerRegisteredIterator

A Iterator implementation that yields EventListenerRegistered signals.

EventListenerRegisteredStream

A Stream implementation that yields EventListenerRegistered signals.

RegistryProxy

RegistryProxyBlocking