Module atspi::socket

DBus interface proxy for: org.a11y.atspi.Socket

This code was generated by zbus-xmlgen 2.0.1 from DBus introspection data. Source: Socket.xml.

You may prefer to adapt it, instead of using it verbatim.

More information can be found in the Writing a client proxy section of the zbus documentation.

Structs

Available

A Available signal.

AvailableArgs

Available signal arguments.

AvailableIterator

A Iterator implementation that yields [signal_name] signals.

AvailableStream

A Stream implementation that yields [signal_name] signals.

SocketProxy

SocketProxyBlocking