atspi

Struct ObjectRef

pub struct ObjectRef {
    pub name: OwnedUniqueName,
    pub path: OwnedObjectPath,
}

ObjectRef type

A ubiquitous type used to refer to an object in the accessibility tree.

In AT-SPI2, objects in the applications’ UI object tree are uniquely identified using a server name and object path. “(so)”

Emitted by RemoveAccessible and Available

Fields
name: OwnedUniqueNamepath: OwnedObjectPath