Struct atspi::StateSet

pub struct StateSet(/* private fields */);

The bitflag representation of all states an object may have.

Implementations

impl StateSet

pub fn new<B: Into<BitFlags<State>>>(value: B) -> Self

Create a new StateSet.

Example

     let states = State::Focusable | State::Sensitive | State::Active;
     let set = StateSet::new(states);

     assert!(set.contains(State::Active));
     assert!(!set.contains(State::Busy));

pub fn from_bits(bits: u64) -> Result<StateSet, FromBitsError<State>>

Returns the StateSet that corresponds to the provided u64s bit pattern.

Errors

When the argument encodes an undefined State.

pub fn empty() -> StateSet

Create an empty StateSet

pub fn bits(&self) -> u64

Returns the state as represented by a u64.

pub fn contains<B: Into<BitFlags<State>>>(self, other: B) -> bool

Whether the StateSet contains a State.

pub fn remove<B: Into<BitFlags<State>>>(&mut self, other: B)

Removes a State (optionally) previously contained in the StateSet.

pub fn insert<B: Into<BitFlags<State>>>(&mut self, other: B)

Inserts a State in the StateSet.

pub fn iter(self) -> impl Iterator<Item = State>

Returns an iterator that yields each set State.

pub fn is_empty(self) -> bool

Checks if all states are unset.

pub fn intersects<B: Into<BitFlags<State>>>(self, other: B) -> bool

Returns true if at least one flag is shared.

pub fn toggle<B: Into<BitFlags<State>>>(&mut self, other: B)

Toggles the matching bits.