Struct atspi::table_cell::TableCellProxyBlocking

pub struct TableCellProxyBlocking<'c>(_);

Implementations

impl<'c> TableCellProxyBlocking<'c>

pub fn new(conn: &Connection) -> Result<TableCellProxyBlocking<'c>>

Creates a new proxy with the default service and path.

pub fn builder(conn: &Connection) -> ProxyBuilder<'c, Self>

Returns a customizable builder for this proxy.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'c>

The reference to the underlying zbus::Proxy.

pub fn get_row_column_span(&self) -> Result<(bool, i32, i32, i32, i32)>

GetRowColumnSpan method

pub fn column_span(&self) -> Result<i32>

ColumnSpan property

pub fn cached_column_span(
&self
) -> Result<Option<<Result<i32> as ResultAdapter>::Ok>, <Result<i32> as ResultAdapter>::Err>

Get the cached value of the ColumnSpan property, or None if the property is not cached.

pub fn receive_column_span_changed(
&self
) -> PropertyIterator<'c, <Result<i32> as ResultAdapter>::Ok>

Create a stream for the ColumnSpan property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn position(&self) -> Result<(i32, i32)>

Position property

pub fn cached_position(
&self
) -> Result<Option<<Result<(i32, i32)> as ResultAdapter>::Ok>, <Result<(i32, i32)> as ResultAdapter>::Err>

Get the cached value of the Position property, or None if the property is not cached.

pub fn receive_position_changed(
&self
) -> PropertyIterator<'c, <Result<(i32, i32)> as ResultAdapter>::Ok>

Create a stream for the Position property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn row_span(&self) -> Result<i32>

RowSpan property

pub fn cached_row_span(
&self
) -> Result<Option<<Result<i32> as ResultAdapter>::Ok>, <Result<i32> as ResultAdapter>::Err>

Get the cached value of the RowSpan property, or None if the property is not cached.

pub fn receive_row_span_changed(
&self
) -> PropertyIterator<'c, <Result<i32> as ResultAdapter>::Ok>

Create a stream for the RowSpan property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn table(&self) -> Result<(String, OwnedObjectPath)>

Table property

pub fn cached_table(
&self
) -> Result<Option<<Result<(String, OwnedObjectPath)> as ResultAdapter>::Ok>, <Result<(String, OwnedObjectPath)> as ResultAdapter>::Err>

Get the cached value of the Table property, or None if the property is not cached.

pub fn receive_table_changed(
&self
) -> PropertyIterator<'c, <Result<(String, OwnedObjectPath)> as ResultAdapter>::Ok>

Create a stream for the Table property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.