Struct atspi::value::ValueProxy

pub struct ValueProxy<'c>(_);

Implementations

impl<'c> ValueProxy<'c>

pub async fn new(conn: &Connection) -> Result<ValueProxy<'c>>

Creates a new proxy with the default service and path.

pub fn builder(conn: &Connection) -> ProxyBuilder<'c, Self>

Returns a customizable builder for this proxy.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'c>

The reference to the underlying zbus::Proxy.

pub async fn current_value(&self) -> Result<f64>

CurrentValue property

pub fn cached_current_value( &self ) -> Result<Option<<Result<f64> as ResultAdapter>::Ok>, <Result<f64> as ResultAdapter>::Err>

Get the cached value of the CurrentValue property, or None if the property is not cached.

pub async fn receive_current_value_changed( &self ) -> PropertyStream<'c, <Result<f64> as ResultAdapter>::Ok>

Create a stream for the CurrentValue property changes. This is a convenient wrapper around zbus::Proxy::receive_property_changed.

pub async fn set_current_value(&self, value: f64) -> Result<()>

Set CurrentValue property

pub async fn maximum_value(&self) -> Result<f64>

MaximumValue property

pub fn cached_maximum_value( &self ) -> Result<Option<<Result<f64> as ResultAdapter>::Ok>, <Result<f64> as ResultAdapter>::Err>

Get the cached value of the MaximumValue property, or None if the property is not cached.

pub async fn receive_maximum_value_changed( &self ) -> PropertyStream<'c, <Result<f64> as ResultAdapter>::Ok>

Create a stream for the MaximumValue property changes. This is a convenient wrapper around zbus::Proxy::receive_property_changed.

pub async fn minimum_increment(&self) -> Result<f64>

MinimumIncrement property

pub fn cached_minimum_increment( &self ) -> Result<Option<<Result<f64> as ResultAdapter>::Ok>, <Result<f64> as ResultAdapter>::Err>

Get the cached value of the MinimumIncrement property, or None if the property is not cached.

pub async fn receive_minimum_increment_changed( &self ) -> PropertyStream<'c, <Result<f64> as ResultAdapter>::Ok>

Create a stream for the MinimumIncrement property changes. This is a convenient wrapper around zbus::Proxy::receive_property_changed.

pub async fn minimum_value(&self) -> Result<f64>

MinimumValue property

pub fn cached_minimum_value( &self ) -> Result<Option<<Result<f64> as ResultAdapter>::Ok>, <Result<f64> as ResultAdapter>::Err>

Get the cached value of the MinimumValue property, or None if the property is not cached.

pub async fn receive_minimum_value_changed( &self ) -> PropertyStream<'c, <Result<f64> as ResultAdapter>::Ok>

Create a stream for the MinimumValue property changes. This is a convenient wrapper around zbus::Proxy::receive_property_changed.