Crate atspi

Modules

cache

DBus interface proxy for: org.a11y.atspi.Cache

error

events

interface

state

Macros

impl_event_conversions

Structs

CacheItem

The item type provided by Cache:Add signals

InterfaceSet

A collection type which encodes the AT-SPI interfaces an accellible object has implemented.

LegacyCacheItem

The item type provided by Cache:Add signals

StateSet

The bitflag representation of all states an object may have.

Enums

AtspiError

An error type that can describe atspi and std and different zbus errors.

ClipType

CoordType

The coordinate type encodes the frame of reference.

Event

Encapsulates the various different accessibility bus signal types.

Granularity

Level of granularity to get text of, in relation to a cursor position.

Interface

AT-SPI interfaces an accessible object can implement.

Layer

Indicates relative stacking order of a atspi_proxies::component::ComponentProxy with respect to the onscreen visual representation of the UI.

MatchType

RelationType

Role

An accessible object role. To think of it in terms of HTML, any semantic element likely has a corollary in this enum. For example: <button>, <input>, <form> or <h4>. Non-semantic elements like <span>, <div> and <b> will not be represented here, and this information is not passed through via the atspi library. TODO: add examples for GTK/Qt libraries in addition to HTML examples.

ScrollType

SortOrder

State

Used by various interfaces indicating every possible state of an accessibility object.

TreeTraversalType

Traits

GenericEvent

Shared behavior of bus Signal events.

Type Aliases

MatchArgs

MatcherArgs

ObjectPair

A pair of (sender, object path with id) which constitutes the fundemental parts of an Accessible object in atspi. NOTE: If you update the name of this type alias, also update the constant in atspi_macros::OBJECT_PAIR_NAME.