Crate atspi

Re-exports

pub use events::Event;

pub use events::EventBody;

pub use error::AtspiError;

pub use zbus;

pub use accessible_id::*;

Modules

accessible

DBus interface proxy for: org.a11y.atspi.Accessible

accessible_ext

accessible_id

action

DBus interface proxy for: org.a11y.atspi.Action

action_ext

application

DBus interface proxy for: org.a11y.atspi.Application

application_ext

bus

DBus interface proxies for: org.a11y.Status, org.a11y.Bus

cache

DBus interface proxy for: org.a11y.atspi.Cache

cache_ext

collection

DBus interface proxy for: org.a11y.atspi.Collection

collection_ext

component

DBus interface proxy for: org.a11y.atspi.Component

component_ext

convertable

device_event_controller

DBus interface proxy for: org.a11y.atspi.DeviceEventController

device_event_controller_ext

device_event_listener

DBus interface proxy for: org.a11y.atspi.DeviceEventListener

device_event_listener_ext

document

DBus interface proxy for: org.a11y.atspi.Document

document_ext

editable_text

DBus interface proxy for: org.a11y.atspi.EditableText

editable_text_ext

error

events

hyperlink

DBus interface proxy for: org.a11y.atspi.Hyperlink

hyperlink_ext

hypertext

DBus interface proxy for: org.a11y.atspi.Hypertext

hypertext_ext

identify

image

DBus interface proxy for: org.a11y.atspi.Image

image_ext

registry

DBus interface proxy for: org.a11y.atspi.Registry

registry_ext

selection

DBus interface proxy for: org.a11y.atspi.Selection

selection_ext

signify

Signified signal types

socket

DBus interface proxy for: org.a11y.atspi.Socket

socket_ext

table

DBus interface proxy for: org.a11y.atspi.Table

table_cell

DBus interface proxy for: org.a11y.atspi.TableCell

table_cell_ext

table_ext

text

DBus interface proxy for: org.a11y.atspi.Text

text_ext

value

DBus interface proxy for: org.a11y.atspi.Value

value_ext

Structs

AccessibilityConnection

A connection to the at-spi bus

InterfaceSet

A collection type which encodes the AT-SPI interfaces an crate::accessible::AccessibleProxy has implemented.

StateSet

The bitflag representation of all states an object may have.

Enums

CoordType

The coordinate type encodes the frame of reference.

Interface

AT-SPI interfaces an crate::accessible::AccessibleProxy can implement.

State

Used by various interfaces indicating every possible state an crate::accessible::AccessibleProxy object can assume.

Traits

AtspiProxy

Functions

set_session_accessibility

Set the IsEnabled property in the session bus.