Crate atspi

Re-exports

pub use events::Event;

pub use events::EventBody;

pub use error::AtspiError;

pub use zbus;

Modules

accessible

DBus interface proxy for: org.a11y.atspi.Accessible

accessible_ext

action

DBus interface proxy for: org.a11y.atspi.Action

application

DBus interface proxy for: org.a11y.atspi.Application

bus

DBus interface proxies for: org.a11y.Status, org.a11y.Bus

cache

DBus interface proxy for: org.a11y.atspi.Cache

collection

DBus interface proxy for: org.a11y.atspi.Collection

component

DBus interface proxy for: org.a11y.atspi.Component

convertable

device_event_controller

DBus interface proxy for: org.a11y.atspi.DeviceEventController

device_event_listener

DBus interface proxy for: org.a11y.atspi.DeviceEventListener

document

DBus interface proxy for: org.a11y.atspi.Document

editable_text

DBus interface proxy for: org.a11y.atspi.EditableText

error

events

hyperlink

DBus interface proxy for: org.a11y.atspi.Hyperlink

hypertext

DBus interface proxy for: org.a11y.atspi.Hypertext

identify

Signified signal types

image

DBus interface proxy for: org.a11y.atspi.Image

processed

This code was generated by zbus-xmlgen 2.0.1 from DBus introspection data. Source: Processed.xml.

registry

DBus interface proxy for: org.a11y.atspi.Registry

selection

DBus interface proxy for: org.a11y.atspi.Selection

socket

DBus interface proxy for: org.a11y.atspi.Socket

table

DBus interface proxy for: org.a11y.atspi.Table

table_cell

DBus interface proxy for: org.a11y.atspi.TableCell

text

DBus interface proxy for: org.a11y.atspi.Text

text_ext

value

DBus interface proxy for: org.a11y.atspi.Value

Structs

Connection

A connection to the at-spi bus

InterfaceSet

A collection type which encodes the AT-SPI interfaces an crate::accessible::AccessibleProxy has implemented.

StateSet

The bitflag representation of all states an object may have.

Enums

CoordType

The coordinate type encodes the frame of reference.

Interface

AT-SPI interfaces an crate::accessible::AccessibleProxy can implement.

State

Used by various interfaces indicating every possible state an crate::accessible::AccessibleProxy object can assume.

Constants

ACCESSIBLE

ATSPI_EVENT

CACHE_ADD

EVENT_LISTENER

QSPI_EVENT

Traits

AtspiProxy

Functions

set_session_accessibility

Set the IsEnabled property in the session bus.