Crate atspi

atspi root crate, which may bring in various subcrates:

atspi_common (required)

atspi_proxies (with use of proxies feature flag)

atspi_connection (with use of connection feature flag)

zbus re-export (with use of zbus feature flag)

Re-exports

pub use atspi_proxies as proxy;

pub use atspi_connection as connection;

Modules

cache

Common types for org.a11y.atspi.Cache events.

error

events

interface

Conversion functions and types representing a set of Interfaces.

object_ref

state

Structs

AccessibilityConnection

A connection to the at-spi bus

CacheItem

The item type provided by Cache:Add signals

InterfaceSet

A collection type which encodes the AT-SPI interfaces an accessible object has implemented.

LegacyCacheItem

The item type provided by Cache:Add signals

ObjectRef

ObjectRef type

StateSet

The bitflag representation of all states an object may have.

Enums

AtspiError

An error type that can describe atspi and std and different zbus errors.

ClipType

Enumeration used by TextProxy to indicate how to treat characters intersecting bounding boxes.

CoordType

The coordinate type encodes the frame of reference.

Event

Encapsulates the various different accessibility bus signal types.

Granularity

Level of granularity to get text of, in relation to a cursor position.

Interface

AT-SPI interfaces an accessible object can implement.

Layer

Indicates relative stacking order of a atspi_proxies::component::ComponentProxy with respect to the onscreen visual representation of the UI.

Live

Enumeration used to indicate a type of live region and how assertive it should be in terms of speaking notifications. Currently, this is only used for Announcement events, but it may be used for additional purposes in the future. The argument in the Announcement event is named politeness.

MatchType

Enumeration used by MatchArgs to specify how to interpret ObjectRef objects.

RelationType

Describes a relationship between one object and another.

Role

An accessible object role. To think of it in terms of HTML, any semantic element likely has a corollary in this enum. For example: <button>, <input>, <form> or <h4>. Non-semantic elements like <span>, <div> and <b> will not be represented here, and this information is not passed through via the atspi library. TODO: add examples for GTK/Qt libraries in addition to HTML examples.

ScrollType

Enumeration used by interface the crate::interface::Interface::Accessible to specify where an object should be placed on the screen when using scroll_to.

SortOrder

Enumeration used by interface CollectionProxy to specify the way ObjectRef objects should be sorted.

State

Used by various interfaces indicating every possible state of an accessibility object.

TreeTraversalType

Method of traversing a tree in the CollectionProxy.

Traits

BusProperties

Describes the DBus-related information about a given struct.

EventProperties

EventProperties allows access to the internals of an event, specifically:

EventTypeProperties

Describes properties of a specific event type.

Type Aliases

MatchArgs

Result