Trait atspi::convertable::ConvertableBlocking

pub trait ConvertableBlocking {

    type Error: Error;
    type Accessible: AccessibleBlocking;
    type Action: ActionBlocking;
    type Application: ApplicationBlocking;
    type Collection: CollectionBlocking;
    type Component: ComponentBlocking;
    type Document: DocumentBlocking;
    type Hypertext: HypertextBlocking;
    type Hyperlink: HyperlinkBlocking;
    type Image: ImageBlocking;
    type Selection: SelectionBlocking;
    type Table: TableBlocking;
    type TableCell: TableCellBlocking;
    type Text: TextBlocking;
    type EditableText: EditableTextBlocking;
    type Value: ValueBlocking;

    fn to_accessible(&self) -> Result<Self::Accessible, Self::Error>;
    fn to_action(&self) -> Result<Self::Action, Self::Error>;
    fn to_application(&self) -> Result<Self::Application, Self::Error>;
    fn to_collection(&self) -> Result<Self::Collection, Self::Error>;
    fn to_component(&self) -> Result<Self::Component, Self::Error>;
    fn to_document(&self) -> Result<Self::Document, Self::Error>;
    fn to_hypertext(&self) -> Result<Self::Hypertext, Self::Error>;
    fn to_hyperlink(&self) -> Result<Self::Hyperlink, Self::Error>;
    fn to_image(&self) -> Result<Self::Image, Self::Error>;
    fn to_selection(&self) -> Result<Self::Selection, Self::Error>;
    fn to_table(&self) -> Result<Self::Table, Self::Error>;
    fn to_table_cell(&self) -> Result<Self::TableCell, Self::Error>;
    fn to_text(&self) -> Result<Self::Text, Self::Error>;
    fn to_editable_text(&self) -> Result<Self::EditableText, Self::Error>;
    fn to_value(&self) -> Result<Self::Value, Self::Error>;

}

Required Associated Types

type Error: Error

type Accessible: AccessibleBlocking

type Action: ActionBlocking

type Application: ApplicationBlocking

type Collection: CollectionBlocking

type Component: ComponentBlocking

type Document: DocumentBlocking

type Hypertext: HypertextBlocking

type Hyperlink: HyperlinkBlocking

type Image: ImageBlocking

type Selection: SelectionBlocking

type Table: TableBlocking

type TableCell: TableCellBlocking

type Text: TextBlocking

type EditableText: EditableTextBlocking

type Value: ValueBlocking

Required Methods

fn to_accessible(&self) -> Result<Self::Accessible, Self::Error>

Creates an Self::Accessible from the existing accessible item.

Errors

This may fail based on the implementation of. Generally, it fails if the accessible item does not implement to accessible interface. This shouldn’t be possible, but this function may fail for other reasons. For example, to convert a zbus::Proxy into a Self::Accessible, it may fail to create the new crate::accessible::AccessibleProxyBlocking.

fn to_action(&self) -> Result<Self::Action, Self::Error>

Creates an Self::Action from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to action interface.

fn to_application(&self) -> Result<Self::Application, Self::Error>

Creates an Self::Application from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to application interface.

fn to_collection(&self) -> Result<Self::Collection, Self::Error>

Creates an Self::Collection from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to collection interface.

fn to_component(&self) -> Result<Self::Component, Self::Error>

Creates an Self::Component from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to component interface.

fn to_document(&self) -> Result<Self::Document, Self::Error>

Creates an Self::Document from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to document interface.

fn to_hypertext(&self) -> Result<Self::Hypertext, Self::Error>

Creates an Self::Hypertext from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to hypertext interface.

fn to_hyperlink(&self) -> Result<Self::Hyperlink, Self::Error>

Creates an Self::Hyperlink from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to hyperlink interface.

fn to_image(&self) -> Result<Self::Image, Self::Error>

Creates an Self::Image from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to image interface.

fn to_selection(&self) -> Result<Self::Selection, Self::Error>

Creates an Self::Selection from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to selection interface.

fn to_table(&self) -> Result<Self::Table, Self::Error>

Creates an Self::Table from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to table interface.

fn to_table_cell(&self) -> Result<Self::TableCell, Self::Error>

Creates an Self::TableCell from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to table cell interface.

fn to_text(&self) -> Result<Self::Text, Self::Error>

Creates an Self::Text from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to text interface.

fn to_editable_text(&self) -> Result<Self::EditableText, Self::Error>

Creates an Self::EditableText from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to editable text interface.

fn to_value(&self) -> Result<Self::Value, Self::Error>

Creates an Self::Value from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to value interface.