atspi::events

Trait BusProperties

pub trait BusProperties {
    const DBUS_MEMBER: &'static str;
    const DBUS_INTERFACE: &'static str;
    const MATCH_RULE_STRING: &'static str;
    const REGISTRY_EVENT_STRING: &'static str;
}

Describes the DBus-related information about a given struct.

DBus member name

DBus interface name

DBus match string: used to tell DBus you are interested in a particular signal

accessibility registry event string: used to tell the accessibility registry that you are interested in a particular event

This trait is not object-safe. For a similar, but object-safe trait, see EventProperties.

Required Associated Constants

const DBUS_MEMBER: &'static str

The DBus member for the event. For example, for an crate::events::object::TextChangedEvent this should be "TextChanged"

const DBUS_INTERFACE: &'static str

The DBus interface name for this event. For example, for any event within crate::events::object, this should be “org.a11y.atspi.Event.Object”.

const MATCH_RULE_STRING: &'static str

A static match rule string for DBus. This should usually be a string that looks like this: "type='signal',interface='org.a11y.atspi.Event.Object',member='PropertyChange'"; This should be deprecated in favour of composing the string from Self::DBUS_MEMBER and Self::DBUS_INTERFACE.

const REGISTRY_EVENT_STRING: &'static str

A registry event string for registering for event receiving via the RegistryProxy. This should be deprecated in favour of composing the string from Self::DBUS_MEMBER and Self::DBUS_INTERFACE.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.